+++ NEWS.md
@@ -1,7 +1,8 @@
 # DedooseR 0.0.4
+* 2025-10-03: Added the compare_saturation function so we can set saturation thresholds and see if codes meet them
+* 2025-10-03: Updated the set_saturation function so we can set by either code counts, transcript coverage or both
 
 # DedooseR 0.0.3
-* 2025-10-03: Updated the set_saturation function so we can set by either code counts, transcript coverage or both
 * 2025-10-03: Updated the plot_counts function so we can plot by either code counts or transcript coverage
 * 2025-10-03: Updated the count_codes function to output transcript coverage